# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: GopherAtlas CMS identity and editorial API
  version: 0.2.0
  description:
    P0-2 private same-origin control plane. Publish selects an immutable SQLite revision,
    claims permanent routes and appends audit in one transaction; it does not rebuild the public
    site. Identity security and the separate Admin-only Monitor are unchanged. JSON requests are
    capped at 1 MiB; Markdown bodies at 512 KiB UTF-8 bytes.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://blog.go-furry.com
    description: Private Tailscale HTTPS origin
  - url: http://127.0.0.1:5173
    description: Loopback Vite same-origin proxy; CMS_BASE_URL/redirect must match
security:
  - Session: []
...
paths:
...
  /api/admin/v1/content:
    get:
      operationId: listContent
      tags:
        - Content
      summary: List visible content
      description:
        Visibility is applied inside the query, before pagination. Only Admin may include
        archived. Reviewer sees others only while in review, with immutable pending titles.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentList'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
      parameters:
        - name: after
          in: query
          description:
            Exclusive keyset cursor; default 0, fixed maximum 100 results. Revision lists use
            revisionNo; other lists use the returned row ID.
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
        - name: includeArchived
          in: query
          description: Admin only when true.
          schema:
            type: boolean
            default: false
    post:
      operationId: createContent
      tags:
        - Content
      summary: Create identity and initial Draft
      description:
        Active roles create own Curated/Post/Note. Topic requires Admin. Initial version 1 may
        be incomplete; creator/owner/byline are the actor. Atomic creation audit.
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
      security:
        - Session: []
          CSRF: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContentInput'
  /api/admin/v1/content/{id}:
    get:
      operationId: getContent
      tags:
        - Content
      summary: Read visible content detail
      description: Reviewer inspecting another owner sees pending immutable material and no mutable Draft.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
      parameters:
        - name: id
          in: path
          required: true
          description: Content or Tag identity.
          schema:
            type: integer
            format: int64
            minimum: 1
  /api/admin/v1/content/{id}/draft:
    put:
      operationId: saveDraft
      tags:
        - Content
      summary: Replace Draft with expected version
      description:
        Owner/Admin only, not archived or in_review. Saving synced changes state to draft and
        preserves published selection. No Audit event.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draft'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
      parameters:
        - name: id
          in: path
          required: true
          description: Content or Tag identity.
          schema:
            type: integer
            format: int64
            minimum: 1
      security:
        - Session: []
          CSRF: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftInput'
  /api/admin/v1/content/{id}/revisions:
    get:
      operationId: listRevisions
      tags:
        - Content
      summary: List immutable revision summaries
      description:
        Reviewer viewing another owner may list only the current pending revision. after is
        exclusive revisionNo.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionList'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
      parameters:
        - name: id
          in: path
          required: true
          description: Content or Tag identity.
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: after
          in: query
          description:
            Exclusive keyset cursor; default 0, fixed maximum 100 results. Revision lists use
            revisionNo; other lists use the returned row ID.
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
  /api/admin/v1/content/{id}/revisions/{revisionNo}:
    get:
      operationId: getRevision
      tags:
        - Content
      summary: Read immutable Revision with relations and review
      description:
        Visibility follows Content policy. Reviewer viewing another owner may fetch only the
        current pending revision.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Revision'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
      parameters:
        - name: id
          in: path
          required: true
          description: Content or Tag identity.
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: revisionNo
          in: path
          required: true
          description: Immutable per-content revision number.
          schema:
            type: integer
            format: int64
            minimum: 1
  /api/admin/v1/content/{id}/revisions/{revisionNo}/actions/restore:
    post:
      operationId: restoreRevision
      tags:
        - Content
      summary: Copy a historical Revision into Draft
      description:
        Owner/Admin, expected current Draft version, not archived/in_review. Copy
        fields/tags/topic order and bump version; state draft. Preserve published pointer and
        immutable history. Non-Admin cannot use restoration to change featured.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
      parameters:
        - name: id
          in: path
          required: true
          description: Content or Tag identity.
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: revisionNo
          in: path
          required: true
          description: Immutable per-content revision number.
          schema:
            type: integer
            format: int64
            minimum: 1
      security:
        - Session: []
          CSRF: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionInput'
...
  /api/admin/v1/content/{id}/actions/request-changes:
    post:
      operationId: requestChanges
      tags:
        - Content
      summary: Request changes on exact pending Revision
      description:
        Reviewer/Admin. Reviewer cannot be owner or revision byline; Admin bypasses. Required
        safe Markdown comment. Append terminal review, clear pending and state changes_requested
        atomically.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
      parameters:
        - name: id
          in: path
          required: true
          description: Content or Tag identity.
          schema:
            type: integer
            format: int64
            minimum: 1
      security:
        - Session: []
          CSRF: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestChangesInput'
...
  /api/admin/v1/reviews:
    get:
      operationId: listReviews
      tags:
        - Reviews
      summary: List review queue or history
      description:
        Reviewer/Admin. Pending queue is immutable revision data and excludes reviewer
        owner/byline self-review. History contains immutable terminal reviews. Each view has its own
        cursor IDs.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/PendingReviewList'
                  - $ref: '#/components/schemas/ReviewHistoryList'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
      parameters:
        - name: after
          in: query
          description:
            Exclusive keyset cursor; default 0, fixed maximum 100 results. Revision lists use
            revisionNo; other lists use the returned row ID.
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
        - name: view
          in: query
          description: Pending immutable submissions or terminal review history.
          schema:
            type: string
            enum:
              - pending
              - history
            default: pending
  /api/admin/v1/tags:
    get:
      operationId: listTags
      tags:
        - Tags
      summary: List selectable tags
      description: All active roles. No hard delete in P0-2.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagList'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
      parameters:
        - name: after
          in: query
          description:
            Exclusive keyset cursor; default 0, fixed maximum 100 results. Revision lists use
            revisionNo; other lists use the returned row ID.
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
    post:
      operationId: createTag
      tags:
        - Tags
      summary: Create tag
      description: Admin only; normalized name and explicit slug must be unique. Atomic audit.
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
      security:
        - Session: []
          CSRF: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagInput'
...
  /api/admin/v1/audit:
    get:
      operationId: listAudit
      tags:
        - Audit
      summary: List safe audit history
      description: Admin/Reviewer only; Editor denied. Draft saves and failed mutations are omitted.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditList'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
      parameters:
        - name: after
          in: query
          description:
            Exclusive keyset cursor; default 0, fixed maximum 100 results. Revision lists use
            revisionNo; other lists use the returned row ID.
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
...
components:
  securitySchemes:
    Session:
      type: apiKey
      in: cookie
      name: __Host-gopheratlas_session
      description:
        'Opaque token; SHA-256 only in SQLite. Secure, HttpOnly, SameSite=Lax, Path=/, no
        Domain. Validated loopback HTTP development uses gopheratlas_dev_session. CSRF cookie:
        __Host-gopheratlas_csrf (or gopheratlas_dev_csrf in HTTP development), non-HttpOnly,
        host-only; mutation headers must match.'
    CSRF:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description:
        Required for mutations. Must match the session-bound same-origin CSRF cookie and stored
        SHA-256 hash. Origin must also exactly match CMS_BASE_URL. Browser transport attaches the
        header automatically.
  responses:
    Error:
      description: Stable error envelope without provider/SQL details. See error.code for classification.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ContentType:
      type: string
      enum:
        - curated_article
        - post
        - note
        - topic
    EditorialState:
      type: string
      enum:
        - draft
        - in_review
        - changes_requested
        - synced
...
    Draft:
      type: object
      additionalProperties: false
      required:
        - version
        - title
        - slug
        - summary
        - bodyMarkdown
        - bylineUserId
        - language
        - featured
        - seoTitle
        - seoDescription
        - payload
        - tagIds
        - topicEntries
        - payloadSchemaVersion
        - updatedBy
        - updatedAt
      properties:
        version:
          type: integer
          format: int64
          minimum: 1
        title:
          type: string
          maxLength: 200
          description: May be empty while authoring; required before submission/publication.
        slug:
          type: string
          maxLength: 100
          pattern: ^(?:[a-z0-9]+(?:-[a-z0-9]+)*)?$
        summary:
          type: string
          maxLength: 4000
        bodyMarkdown:
          type: string
          description:
            At most 524288 UTF-8 bytes. Server CommonMark/GFM safety rules reject H1, frontmatter,
            raw HTML/MDX, unsafe URLs and uncontrolled images.
        bylineUserId:
          type: integer
          format: int64
          minimum: 1
        language:
          type: string
          maxLength: 32
          description: Empty while authoring, otherwise bounded language tag; required before submission.
        featured:
          type: boolean
        seoTitle:
          type: string
          maxLength: 120
        seoDescription:
          type: string
          maxLength: 320
        payload:
          $ref: '#/components/schemas/TypedPayload'
        tagIds:
          type: array
          items:
            type: integer
            format: int64
            minimum: 1
          maxItems: 100
          uniqueItems: true
        topicEntries:
          type: array
          items:
            $ref: '#/components/schemas/TopicEntry'
          maxItems: 100
        payloadSchemaVersion:
          type: integer
          const: 1
        updatedBy:
          type: integer
          format: int64
          minimum: 1
        updatedAt:
          type: integer
          format: int64
          minimum: 0
...
    Content:
      type: object
      additionalProperties: false
      required:
        - id
        - type
        - ownerUserId
        - editorialState
        - pendingReviewRevisionId
        - publishedRevisionId
        - title
        - createdBy
        - createdAt
        - updatedAt
        - firstPublishedAt
        - lastPublishedAt
        - archivedAt
        - draft
        - pendingRevision
        - routes
        - nextRouteCursor
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        type:
          $ref: '#/components/schemas/ContentType'
        ownerUserId:
          type: integer
          format: int64
          minimum: 1
        editorialState:
          $ref: '#/components/schemas/EditorialState'
        pendingReviewRevisionId:
          type:
            - integer
            - 'null'
          format: int64
        publishedRevisionId:
          type:
            - integer
            - 'null'
          format: int64
        title:
          type: string
          maxLength: 200
        createdBy:
          type: integer
          format: int64
          minimum: 1
        createdAt:
          type: integer
          format: int64
          minimum: 0
        updatedAt:
          type: integer
          format: int64
          minimum: 0
        firstPublishedAt:
          type:
            - integer
            - 'null'
          format: int64
        lastPublishedAt:
          type:
            - integer
            - 'null'
          format: int64
        archivedAt:
          type:
            - integer
            - 'null'
          format: int64
        draft:
          anyOf:
            - $ref: '#/components/schemas/Draft'
            - type: 'null'
        pendingRevision:
          anyOf:
            - $ref: '#/components/schemas/Revision'
            - type: 'null'
        routes:
          type: array
          items:
            $ref: '#/components/schemas/Route'
          maxItems: 100
        nextRouteCursor:
          type:
            - integer
            - 'null'
          format: int64
          description: Exclusive cursor for GET /content/{id}/routes when more route history may exist.
      description:
        Editors see their own content. Reviewer access to another owner is limited to an
        in-review immutable pending revision; draft is null. Admin sees all, including archived. A
        completed action may return summary with null draft/pendingRevision when the actor no longer
        has view access.
...
    ContentList:
      type: object
      additionalProperties: false
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ContentSummary'
          maxItems: 100
        nextCursor:
          type:
            - integer
            - 'null'
          format: int64
          description:
            Use as after for the next page. Null ends pagination. A full final page may yield one
            empty next page.
    RevisionList:
      type: object
      additionalProperties: false
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RevisionSummary'
          maxItems: 100
        nextCursor:
          type:
            - integer
            - 'null'
          format: int64
          description:
            Use as after for the next page. Null ends pagination. A full final page may yield one
            empty next page.
    PendingReviewList:
      type: object
      additionalProperties: false
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PendingReview'
          maxItems: 100
        nextCursor:
          type:
            - integer
            - 'null'
          format: int64
          description:
            Use as after for the next page. Null ends pagination. A full final page may yield one
            empty next page.
...
    CreateContentInput:
      type: object
      additionalProperties: false
      required:
        - type
      properties:
        type:
          $ref: '#/components/schemas/ContentType'
    VersionInput:
      type: object
      additionalProperties: false
      required:
        - version
      properties:
        version:
          type: integer
          format: int64
          minimum: 1
    RequestChangesInput:
      type: object
      additionalProperties: false
      required:
        - revisionId
        - commentMarkdown
      properties:
        revisionId:
          type: integer
          format: int64
          minimum: 1
        commentMarkdown:
          type: string
          description: Required nonblank safe Markdown, at most 16384 UTF-8 bytes.
          minLength: 1
...